Made UtilizationDiagram responsive, add JavaDoc
Remove the 'clip' property (which does not allow percent) from the 
.quarters CSS class and make every quarter to a square that is half the
width and height of its parent, has a border-top-right-radius of 100%
and rotates around its bottom right corner.

Change rendering of the circle:
Move the fourth quarter instead of the final quarter for a percentage
between 70 and 100 and set the rotation of the final quarter about 270
degrees.